Issue #1336: Relocate files to new packages
