+++ app/src/androidTest/java/org/mozilla/tv/firefox/screenshots/TVScreenshots.java
@@ -42,7 +42,7 @@ import static android.support.test.espresso.matcher.ViewMatchers.withText;
 import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.core.IsNot.not;
-import static org.mozilla.tv.firefox.OnboardingActivity.ONBOARD_SHOWN_PREF;
+import static org.mozilla.tv.firefox.onboarding.OnboardingActivity.ONBOARD_SHOWN_PREF;
 import static org.mozilla.tv.firefox.pocket.PocketOnboardingActivity.POCKET_ONBOARDING_SHOWN_PREF;
