+++ app/src/main/java/org/mozilla/tv/firefox/components/locale/LocaleAwareFragment.java
@@ -3,7 +3,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 
-package org.mozilla.tv.firefox.locale;
+package org.mozilla.tv.firefox.components.locale;
 
 import android.support.v4.app.Fragment;
