+++ app/src/main/java/org/mozilla/tv/firefox/components/search/SearchEngineParser.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.tv.firefox.search;
+package org.mozilla.tv.firefox.components.search;
 
 import android.content.res.AssetManager;
 import android.graphics.BitmapFactory;
