+++ app/src/main/java/org/mozilla/tv/firefox/onboarding/OnboardingActivity.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.tv.firefox
+package org.mozilla.tv.firefox.onboarding
 
 import android.app.Activity
 import android.content.Intent
@@ -10,6 +10,7 @@ import android.os.Bundle
 import android.preference.PreferenceManager
 import android.support.v7.app.AppCompatActivity
 import kotlinx.android.synthetic.main.content_onboarding.*
+import org.mozilla.tv.firefox.R
 import org.mozilla.tv.firefox.utils.Settings
 
 class OnboardingActivity : AppCompatActivity() {
