+++ app/src/main/java/org/mozilla/tv/firefox/telemetry/TelemetryIntegration.kt
@@ -10,12 +10,12 @@ import android.net.http.SslError
 import android.os.StrictMode
 import android.support.annotation.AnyThread
 import android.support.annotation.UiThread
-import org.mozilla.tv.firefox.webrender.NavigationEvent
+import org.mozilla.tv.firefox.navigationoverlay.NavigationEvent
 import org.mozilla.tv.firefox.ext.resetAfter
 import org.mozilla.tv.firefox.pinnedtile.BundledHomeTile
 import org.mozilla.tv.firefox.pinnedtile.CustomHomeTile
 import org.mozilla.tv.firefox.pinnedtile.HomeTile
-import org.mozilla.tv.firefox.search.SearchEngineManager
+import org.mozilla.tv.firefox.components.search.SearchEngineManager
 import org.mozilla.tv.firefox.utils.Assert
 import org.mozilla.tv.firefox.widget.InlineAutocompleteEditText.AutocompleteResult
 import org.mozilla.telemetry.TelemetryHolder
