+++ app/src/main/java/org/mozilla/tv/firefox/webrender/Components.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.tv.firefox
+package org.mozilla.tv.firefox.webrender
 
 import android.content.Context
 import mozilla.components.browser.engine.system.SystemEngine
@@ -10,8 +10,7 @@ import mozilla.components.browser.session.SessionManager
 import mozilla.components.concept.engine.DefaultSettings
 import mozilla.components.concept.engine.Engine
 import mozilla.components.feature.session.SessionUseCases
-import org.mozilla.tv.firefox.webrender.UserAgent
-import org.mozilla.tv.firefox.engine.CustomContentRequestInterceptor
+import org.mozilla.tv.firefox.R
 import org.mozilla.tv.firefox.utils.Settings
 
 /**
