+++ app/src/main/res/layout/browser_overlay.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?><!-- This Source Code Form is subject to the terms of the Mozilla Public
    - License, v. 2.0. If a copy of the MPL was not distributed with this
    - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
-<org.mozilla.tv.firefox.webrender.BrowserNavigationOverlayScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+<org.mozilla.tv.firefox.navigationoverlay.BrowserNavigationOverlayScrollView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/overlayScrollView"
     android:layout_width="match_parent"
@@ -91,4 +91,4 @@
 
         </LinearLayout>
     </android.support.v4.widget.NestedScrollView>
-</org.mozilla.tv.firefox.webrender.BrowserNavigationOverlayScrollView>
+</org.mozilla.tv.firefox.navigationoverlay.BrowserNavigationOverlayScrollView>
