+++ app/src/main/res/layout/content_onboarding.xml
@@ -5,7 +5,7 @@
     android:layout_width="fill_parent"
     android:layout_height="fill_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="org.mozilla.tv.firefox.OnboardingActivity"
+    tools:context="org.mozilla.tv.firefox.onboarding.OnboardingActivity"
     tools:showIn="@layout/activity_onboarding">
 
     <android.support.constraint.Guideline
