+++ app/src/test/java/org/mozilla/tv/firefox/components/locale/LocalesTest.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.tv.firefox.locale;
+package org.mozilla.tv.firefox.components.locale;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
