+++ app/src/test/java/org/mozilla/tv/firefox/engine/CustomContentRequestInterceptorTest.kt
@@ -14,6 +14,7 @@ import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mockito.mock
+import org.mozilla.tv.firefox.webrender.CustomContentRequestInterceptor
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.RuntimeEnvironment
